Move into separate __tests__ folders
