+++ frontend/app/composables/__tests__/use-passwords.test.ts
@@ -1,6 +1,6 @@
 import { ref } from "vue";
 import { describe, expect, test } from "vitest";
-import { usePasswordStrength } from "./use-passwords";
+import { usePasswordStrength } from "../use-passwords";
 import { stubI18n } from "~/tests/utils";
 
 describe("test usePasswordStrength", () => {
